+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This file contains the list of changes made to pyjoulescope.
 
 *   Fixed PatternBuffer.__len__, which was breaking the unit test in Python 3.8.
 *   Eliminated None control transfer reference when closing device under Windows.
+*   Improved statistics entry point to display both statistics sources.
 
 
 ## 0.9.0